+++ Source/OxyPlot/Series/BoxPlotItem.cs
@@ -9,6 +9,7 @@
 
 namespace OxyPlot.Series
 {
+    using System;
     using System.Collections.Generic;
 
     /// <summary>
@@ -27,6 +28,7 @@ namespace OxyPlot.Series
         /// <param name="upperWhisker">The upper whisker.</param>
         /// <param name="outliers">The outliers.</param>
         /// <param name="tag">The tag.</param>
+        [Obsolete("Will be removed in a future version. Please use the other constructor in combination with an Object Initzializer.")]
         public BoxPlotItem(
             double x,
             double lowerWhisker,
@@ -34,7 +36,7 @@ namespace OxyPlot.Series
             double median,
             double boxTop,
             double upperWhisker,
-            IList<double> outliers = null,
+            IList<double> outliers,
             object tag = null)
             : this()
         {
@@ -44,8 +46,38 @@ namespace OxyPlot.Series
             this.Median = median;
             this.BoxTop = boxTop;
             this.UpperWhisker = upperWhisker;
+            this.Mean = double.NaN;
             this.Outliers = outliers ?? new List<double>();
             this.Tag = tag;
+        }
+
+        /// <summary>
+        /// Initializes a new instance of the <see cref="BoxPlotItem" /> struct.
+        /// </summary>
+        /// <param name="x">The x.</param>
+        /// <param name="lowerWhisker">The lower whisker.</param>
+        /// <param name="boxBottom">The box bottom.</param>
+        /// <param name="median">The median.</param>
+        /// <param name="boxTop">The box top.</param>
+        /// <param name="upperWhisker">The upper whisker.</param>
+        public BoxPlotItem(
+            double x,
+            double lowerWhisker,
+            double boxBottom,
+            double median,
+            double boxTop,
+            double upperWhisker)
+            : this()
+        {
+            this.X = x;
+            this.LowerWhisker = lowerWhisker;
+            this.BoxBottom = boxBottom;
+            this.Median = median;
+            this.BoxTop = boxTop;
+            this.UpperWhisker = upperWhisker;
+            this.Mean = double.NaN;
+            this.Outliers = new List<double>();
+            this.Tag = null;
         }
 
         /// <summary>
@@ -71,6 +103,12 @@ namespace OxyPlot.Series
         /// </summary>
         /// <value>The median.</value>
         public double Median { get; set; }
+
+        /// <summary>
+        /// Gets or sets the mean.
+        /// </summary>
+        /// <value>The mean.</value>
+        public double Mean { get; set; }
 
         /// <summary>
         /// Gets or sets the outliers.
@@ -99,6 +137,12 @@ namespace OxyPlot.Series
             {
                 var values = new List<double> { this.LowerWhisker, this.BoxBottom, this.Median, this.BoxTop, this.UpperWhisker };
 
+                // As mean is an optional value and should not be checked for validation if not set don't add it if NaN
+                if (!double.IsNaN(this.Mean))
+                {
+                    values.Add(this.Mean);
+                }
+
                 values.AddRange(this.Outliers);
 
                 return values;
@@ -118,11 +162,12 @@ namespace OxyPlot.Series
         public override string ToString()
         {
             return string.Format(
-                "{0} {1} {2} {3} {4} {5} ",
+                "{0} {1} {2} {3} {4} {5} {6} ",
                 this.X,
                 this.LowerWhisker,
                 this.BoxBottom,
                 this.Median,
+                this.Mean,
                 this.BoxTop,
                 this.UpperWhisker);
         }